requirements.txt
addict==2.4.0
affine==2.4.0
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
appdirs==1.4.4
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click-plugins==1.1.1
click==8.1.7
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.4
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
fonttools==4.43.1
ftfy==6.1.1
gitdb==4.0.11
gitpython==3.1.40
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy-loader==0.3
markdown-it-py==3.0.0
markdown==3.5
mat4py==0.6.0
matplotlib==3.8.0
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.9.0
mmpretrain==1.1.0
mmsegmentation==1.2.1
model-index==0.1.11
modelindex==0.0.2
networkx==3.2
numpy==1.26.1
opencv-python-headless==4.8.1.78
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.1.1
pathtools==0.1.2
pillow==10.0.1
pip==23.3
platformdirs==3.11.0
prettytable==3.9.0
protobuf==4.24.4
psutil==5.9.6
pycparser==2.21
pycryptodome==3.19.0
pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
qudida==0.0.4
rasterio==1.3.9
regex==2023.10.3
requests==2.28.2
rich==13.4.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==60.2.0
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.65.2
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wandb==0.15.12
wcwidth==0.2.8
wheel==0.41.2
yapf==0.40.2
zipp==3.17.0